+++ R/geom_pointpath.R
@@ -32,6 +32,14 @@
 #' @examples
 #' ggplot(pressure, aes(temperature, pressure)) +
 #'   geom_pointpath()
+#'
+#' # Using geom_pointpath as annotation
+#' ggplot() +
+#'   annotate(
+#'     "pointpath",
+#'     x = c(1, 0.32, 0.31, -0.12, -0.81, -0.4, -0.81, -0.12, 0.31, 0.32, 1),
+#'     y = c(0, 0.24, 0.95, 0.38, 0.59, 0, -0.59, -0.38, -0.95, -0.24, 0)
+#'   )
 geom_pointpath <- function(
   mapping = NULL, data = NULL, stat = "identity",
   position = "identity", ..., na.rm = FALSE, show.legend = NA,
@@ -123,6 +131,12 @@ GeomPointPath <- ggplot2::ggproto(
   ),
   non_missing_aes = c("size", "colour")
 )
+
+#' @usage NULL
+#' @format NULL
+#' @export
+#' @rdname ggh4x_extensions
+GeomPointpath <- GeomPointPath
 
 # Draw methods ----------------------------------------