+++ R/guide_axis_logticks.R
@@ -144,8 +144,10 @@ guide_train.axis_logticks <- function(guide, scale, aesthetic = NULL) {
   )
 
   if (length(intersect(scale$aesthetics, guide$available_aes)) == 0) {
-    warning("axis_logticks guide needs appropriate scales: ",
-            guide$available_aes)
+    cli::cli_warn(c(
+      "{.fn guide_axis_logticks} needs appropriate scales.",
+      i = "Use one of {.or {.field {guide$available_aes}}}."
+    ))
     guide$key <- empty_ticks
   } else if (length(breaks) == 0) {
     guide$key <- empty_ticks
@@ -165,8 +167,7 @@ guide_train.axis_logticks <- function(guide, scale, aesthetic = NULL) {
     guide$key <- ticks[is.finite(ticks[[aesthetic]]), ]
   }
   guide$name <- paste0(guide$name, "_", aesthetic)
-  guide$hash <- digest::digest(list(guide$title, guide$key$.value,
-                                    guide$key$.label, guide$name))
+  guide$hash <- with(guide, hash(list(title, key$.value, key$.label, name)))
   guide <- truncate_guide(guide, scale, aesthetic)
   guide
 }