+++ R/guide_axis_minor.R
@@ -74,7 +74,6 @@ guide_axis_minor <- function(
 guide_train.axis_minor <- function(
   guide, scale, aesthetic = NULL
 ) {
-  try_require("digest", "guide_axis_minor")
   aesthetic <- aesthetic %||% scale$aesthetics[1]
   # Get major and minor breaks
   breaks_major <- scale$get_breaks()
@@ -89,8 +88,10 @@ guide_train.axis_minor <- function(
   )
   names(empty_ticks)[1] <- aesthetic
   if (length(intersect(scale$aesthetics, guide$available_aes)) == 0) {
-    warning("axis_minor guide needs appropriate scales: ",
-            guide$available_aes)
+    cli::cli_warn(c(
+      "{.fn guide_axis_minor} needs appropriate scales.",
+      i = "Use one of {.or {.field {guide$available_aes}}}."
+    ))
     guide$key <- empty_ticks
   } else if (length(breaks) == 0) {
     guide$key <- empty_ticks
@@ -110,8 +111,7 @@ guide_train.axis_minor <- function(
     guide$key <- ticks[is.finite(ticks[[aesthetic]]), ]
   }
   guide$name <- paste0(guide$name, "_", aesthetic)
-  guide$hash <- digest::digest(list(guide$title, guide$key$.value,
-                                    guide$key$.label, guide$name))
+  guide$hash <- with(guide, hash(list(title, key$.value, key$.label, name)))
   guide <- truncate_guide(guide, scale, aesthetic)
   guide
 }