+++ R/guide_axis_truncated.R
@@ -238,8 +238,10 @@ truncate_guide <- function(guide, scale, aesthetic) {
     guide$key[[aesthetic]], guide$trunc_upper, scale, "upper"
   )
   if (!(length(guide$trunc_lower) == length(guide$trunc_upper))) {
-    abort(paste0("Axis truncation must have an equal number of upper and lower",
-                 " truncation points."))
+    cli::cli_abort(paste0(
+      "Axis truncation must have an equal number of upper and lower ",
+      "truncation points."
+    ))
   }
   trunc <- data_frame0(
     "start" = trunc_lower,
@@ -289,8 +291,10 @@ check_trunc_arg <- function(lower, upper) {
   if (!is.function(lower) && !is.function(upper)) {
     lens <- c(length(lower), length(upper))
     if (lens[1] != lens[2] & !any(lens == 0)) {
-      abort(paste0("Axis truncation must have an equal number of upper and ",
-                   "lower truncation points."))
+      cli::cli_abort(paste0(
+        "Axis truncation must have an equal number of upper and lower ",
+        "truncation points."
+      ))
     }
   }
 }