+++ tests/testthat/test-guide_axis_minor.R
@@ -86,5 +86,5 @@ test_that("guide_axis_minor errors upon misuse", {
   gui$available_aes <- "z"
 
   g <- base + scale_x_continuous(guide = gui)
-  expect_warning(ggplotGrob(g), "axis_minor guide needs appropriate scales: z")
+  expect_snapshot_warning(ggplotGrob(g))
 })